Add mouse sensitivity setting, fix WebSocket frame handling and mouse overflow
- Add configurable mouse speed (Slow/Normal/Fast/Very Fast) to web UI with
  localStorage persistence
- Replace mouse FIFO queue with accumulator pattern that sums int16 deltas
  and drains ±127 per HID report, preventing queue overflow
- Add WebSocket frame reassembly buffer to handle TCP-coalesced frames that
  were previously silently dropped
- Fix use-after-free crash when WebSocket close frame was processed during
  frame reassembly (tcp_recved on freed PCB)
- Widen mouse values from int8 to int16 across WebSocket and MQTT handlers
- Fix hidgui mouse grab (warp-to-center) and int8 overflow on fast movement

+++ src/usb.h
@@ -39,7 +39,7 @@ void tud_suspend_cb(bool remote_wakeup_en);
 void tud_resume_cb(void);
 void press_key(uint16_t key);
 void depress_key(uint16_t key);
-void move_mouse(uint8_t buttons, int8_t x, int8_t y, int8_t vertical, int8_t horizontal);
+void move_mouse(uint8_t buttons, int16_t x, int16_t y, int16_t vertical, int16_t horizontal);
 void press_consumer(uint16_t code);
 void release_consumer(void);
 void press_system(uint16_t code);
